Database.Translation

Documentation

type WordAndLang = (Word, Language)

type WordDescription = (Entity Word, [Language], [WordTranslation], [WordSource])

type WordSource = (WordAndLang, [WordTranslation])

type WordTranslation = (Translation, Language, Maybe Word)

addTranslationFromAlt

Arguments

:: (MonadIO m, MonadLogger m)

=> WordText
Text of source word

-> PartOfSpeech
PartOfSpeech of source word

-> LanguageName
LanguageName of source word

-> LanguageName
LanguageName of translation word

-> Text
Text of the translation

-> Maybe Comment
Maybe Text of translation comment

-> AppT m (Maybe (Key Translation))

The addTranslationFromAlt function inserts new translation not to a word.

>>> runSQLAction $ addTranslationFromAlt "bhrentos" Noun Titan English "animal horned" Nothing
Just (TranslationKey {unTranslationKey = SqlBackendKey {unSqlBackendKey = 19980}})

addTranslationFromTo

Arguments

:: (MonadIO m, MonadLogger m)

=> WordText
Text of source word

-> PartOfSpeech
PartOfSpeech of source word

-> LanguageName
LanguageName of source word

-> WordText
Text of translation word

-> PartOfSpeech
PartOfSpeech of translation word

-> LanguageName
LanguageName of translation word

-> Maybe Comment
Maybe Text of translation comment

-> AppT m (Maybe (Key Translation))
Key Translation if inserted succsesfully

The addTranslationFromTo function inserts new translation to a existing word.

>>> runSQLAction $ addTranslationFromTo "kibil" Noun Khuzdûl "silver" Noun English (Just "metal")
Just (TranslationKey {unTranslationKey = SqlBackendKey {unSqlBackendKey = 19980}})

getFullWordDescription :: MonadIO m => [Entity Word] -> AppT m [WordDescription]

translateWord :: MonadIO m => Text -> AppT m [WordDescription]